Fix modules.deps handling
Don't ignore error when modules.deps cannot be included. make failed silently
if modules.deps cannot be generated (due to an invalid module name for
instance).

modules.deps now depends on the top-level Makefile and will be regenerated when
MODULES is updated, as it should be (not need to manually clean).