bootloader: use ROME for communication
On XMEGA, use a specific implementation (don't reuse module)

Two built-in ROME messages are defined for bootloader. A subcommand
field is used to allow a large range of bootloader commands without
using a lot of ROME message IDs.

Since data is checked using ROME's CRCs, there is no need for a CRC on
'prog_page' and 'prog_user_sig' commands. Also, since the client decides
when all the data has been sent, there is no need for a 'size' parameter
for 'prog_user_sig' command.

Page data (application and signature) is chunked and sent using several
'buffer' commands.

Compile with -fno-jump-tables because of this bug (fixed on gcc 4.9):
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=63223