rome: detect when rome_messages module changed
When ROME_MESSAGES is not defined, the default rome_messages module is
used. In such case, detect when its source file changed and trigger
generation.

It assumes that the file rome_messages module is a .py or .pyc file.
Also, if the module is not found, an (explicit) Python traceback is
output.