Group lib/common symbols at the beginning of binaries
Use a linker script and custom sections to order symbols in the
generated binaries. Putting common code at the beginning and specific
code at the end will reduce the diff between sucessively compiled
binaries and thus allow for faster bootloader uploads.

.progmem end is aligned on a 4k to avoid shifting the whole .text when a
string literal (typically, a log) is added.

Since we use -fdata-sections and -ffunctions-sections, we cannot use
objcopy's --rename-sections. We use a section prefix instead.

Note that for XMEGA, the same linker script can be used for all devices.
OUTPUT_ARCH() is deduced and don't have to be enforced in the script,
This was the only line to differ in default scripts.